Improve accuracy of symbol frame rendering

+++ Core/AudioGeneration/IAudioScaler.cs
@@ -3,6 +3,6 @@
     public interface IAudioScaler
     {
         float[] Scale(float[] samples, int sampleRate, int baudRate, int numberOfSymbols, int scaleWidth, int scaleHeight);
-        int SamplesPerSymbol { get; set; }
+        double SamplesPerSymbol { get; set; }
     }
 }
